use css-modules in ArticleCard.tsx
